pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
    "Cython>=3.0.0",
    "numpy>=1.20.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "seam-carver"
version = "0.1.0"
authors = [ 
    {name="Wren C. Kuratana", email="wren@wrenckuratana.com"}, 
    {name="Machiavelli Merkle-Ward", email="mmerklew@macalester.edu"} 
]
description = "A high-performance C/Python seam carving project."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy",
    "Pillow",
    "typer",
    "librosa",
    "soundfile"
]

[project.scripts]
seam-carver = "seam_carver.cli:main"